+++ comps/image2image/sdwebui/image2image_sdwebui.py
@@ -8,11 +8,12 @@ import threading
 import time
 
 import torch
-from diffusers import DiffusionPipeline
+from diffusers import AutoPipelineForImage2Image
+from diffusers.utils import load_image
 
 from comps import (
     CustomLogger,
-    SDInputs,
+    SDImg2ImgInputs,
     SDOutputs,
     ServiceType,
     opea_microservices,
@@ -21,7 +22,7 @@ from comps import (
     statistics_dict,
 )
 
-logger = CustomLogger("text2image_sdwebui")
+logger = CustomLogger("image2image_sdwebui")
 pipe = None
 args = None
 initialization_lock = threading.Lock()
@@ -49,27 +50,19 @@ def initialize():
                         "token": args.token,
                     }
                 )
-                if "stable-diffusion-3" in args.model_name_or_path:
-                    from optimum.habana.diffusers import GaudiStableDiffusion3Pipeline
+                if "stable-diffusion-xl" in args.model_name_or_path:
+                    from optimum.habana.diffusers import GaudiStableDiffusionXLImg2ImgPipeline
 
-                    pipe = GaudiStableDiffusion3Pipeline.from_pretrained(
-                        args.model_name_or_path,
-                        **kwargs,
-                    )
-                elif "stable-diffusion" in args.model_name_or_path.lower() or "flux" in args.model_name_or_path.lower():
-                    from optimum.habana.diffusers import AutoPipelineForText2Image
-
-                    pipe = AutoPipelineForText2Image.from_pretrained(
+                    pipe = GaudiStableDiffusionXLImg2ImgPipeline.from_pretrained(
                         args.model_name_or_path,
                         **kwargs,
                     )
                 else:
                     raise NotImplementedError(
-                        "Only support stable-diffusion, stable-diffusion-xl, stable-diffusion-3 and flux now, "
-                        + f"model {args.model_name_or_path} not supported."
+                        "Only support stable-diffusion-xl now, " + f"model {args.model_name_or_path} not supported."
                     )
             elif args.device == "cpu":
-                pipe = DiffusionPipeline.from_pretrained(args.model_name_or_path, token=args.token, **kwargs)
+                pipe = AutoPipelineForImage2Image.from_pretrained(args.model_name_or_path, token=args.token, **kwargs)
             else:
                 raise NotImplementedError(f"Only support cpu and hpu device now, device {args.device} not supported.")
             logger.info("Stable Diffusion model initialized.")
@@ -77,18 +70,19 @@ def initialize():
 
 
 @register_microservice(
-    name="opea_service@text2image_sdwebui",
-    service_type=ServiceType.TEXT2IMAGE,
-    endpoint="/sdapi/v1/txt2img",
+    name="opea_service@image2image_sdwebui",
+    service_type=ServiceType.IMAGE2IMAGE,
+    endpoint="/sdapi/v1/img2img",
     host="0.0.0.0",
-    port=9379,
-    input_datatype=SDInputs,
+    port=9389,
+    input_datatype=SDImg2ImgInputs,
     output_datatype=SDOutputs,
 )
-@register_statistics(names=["opea_service@text2image_sdwebui"])
-def text2image(input: SDInputs):
+@register_statistics(names=["opea_service@image2image_sdwebui"])
+def image2image(input: SDImg2ImgInputs):
     initialize()
     start = time.time()
+    image = load_image(input.image).convert("RGB")
     prompt = input.prompt
     num_inference_steps = input.num_inference_steps
     guidance_scale = input.guidance_scale
@@ -97,12 +91,26 @@ def text2image(input: SDInputs):
     negative_prompt = input.negative_prompt
     height = input.height
     width = input.width
+    strength = input.strength
 
-    generator = torch.manual_seed(seed)
     images = pipe(
         prompt,
         height=height,
         width=width,
+        num_inference_steps=num_inference_steps,
+        guidance_scale=guidance_scale,
+        generator=generator,
+        negative_prompt=negative_prompt,
+        num_images_per_prompt=num_images_per_prompt,
+    ).images
+
+    generator = torch.manual_seed(seed)
+    images = pipe(
+        image=image,
+        prompt=prompt,
+        height=height,
+        width=width,
+        strength=strength,
         num_inference_steps=num_inference_steps,
         guidance_scale=guidance_scale,
         generator=generator,
@@ -119,13 +127,13 @@ def text2image(input: SDInputs):
             bytes = f.read()
         b64_str = base64.b64encode(bytes).decode()
         results.append(b64_str)
-    statistics_dict["opea_service@text2image_sdwebui"].append_latency(time.time() - start, None)
+    statistics_dict["opea_service@image2image_sdwebui"].append_latency(time.time() - start, None)
     return SDOutputs(images=results)
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("--model_name_or_path", type=str, default="stabilityai/stable-diffusion-3-medium-diffusers")
+    parser.add_argument("--model_name_or_path", type=str, default="stable-diffusion-v1-5/stable-diffusion-v1-5")
     parser.add_argument("--use_hpu_graphs", default=False, action="store_true")
     parser.add_argument("--device", type=str, default="cpu")
     parser.add_argument("--token", type=str, default=None)
@@ -133,5 +141,5 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
 
-    logger.info("Text2image server started.")
-    opea_microservices["opea_service@text2image_sdwebui"].start()
+    logger.info("Image2image server started.")
+    opea_microservices["opea_service@image2image_sdwebui"].start()
