+++ comps/text2image/sdwebui/text2image_sdwebui.py
@@ -8,12 +8,11 @@ import threading
 import time
 
 import torch
-from diffusers import AutoPipelineForImage2Image
-from diffusers.utils import load_image
+from diffusers import DiffusionPipeline
 
 from comps import (
     CustomLogger,
-    SDImg2ImgInputs,
+    SDInputs,
     SDOutputs,
     ServiceType,
     opea_microservices,
@@ -22,7 +21,7 @@ from comps import (
     statistics_dict,
 )
 
-logger = CustomLogger("image2image_sdwebui")
+logger = CustomLogger("text2image_sdwebui")
 pipe = None
 args = None
 initialization_lock = threading.Lock()
@@ -50,19 +49,27 @@ def initialize():
                         "token": args.token,
                     }
                 )
-                if "stable-diffusion-xl" in args.model_name_or_path:
-                    from optimum.habana.diffusers import GaudiStableDiffusionXLImg2ImgPipeline
+                if "stable-diffusion-3" in args.model_name_or_path:
+                    from optimum.habana.diffusers import GaudiStableDiffusion3Pipeline
 
-                    pipe = GaudiStableDiffusionXLImg2ImgPipeline.from_pretrained(
+                    pipe = GaudiStableDiffusion3Pipeline.from_pretrained(
+                        args.model_name_or_path,
+                        **kwargs,
+                    )
+                elif "stable-diffusion" in args.model_name_or_path.lower() or "flux" in args.model_name_or_path.lower():
+                    from optimum.habana.diffusers import AutoPipelineForText2Image
+
+                    pipe = AutoPipelineForText2Image.from_pretrained(
                         args.model_name_or_path,
                         **kwargs,
                     )
                 else:
                     raise NotImplementedError(
-                        "Only support stable-diffusion-xl now, " + f"model {args.model_name_or_path} not supported."
+                        "Only support stable-diffusion, stable-diffusion-xl, stable-diffusion-3 and flux now, "
+                        + f"model {args.model_name_or_path} not supported."
                     )
             elif args.device == "cpu":
-                pipe = AutoPipelineForImage2Image.from_pretrained(args.model_name_or_path, token=args.token, **kwargs)
+                pipe = DiffusionPipeline.from_pretrained(args.model_name_or_path, token=args.token, **kwargs)
             else:
                 raise NotImplementedError(f"Only support cpu and hpu device now, device {args.device} not supported.")
             logger.info("Stable Diffusion model initialized.")
@@ -70,19 +77,18 @@ def initialize():
 
 
 @register_microservice(
-    name="opea_service@image2image_sdwebui",
-    service_type=ServiceType.IMAGE2IMAGE,
-    endpoint="/sdapi/v1/img2img",
+    name="opea_service@text2image_sdwebui",
+    service_type=ServiceType.TEXT2IMAGE,
+    endpoint="/sdapi/v1/txt2img",
     host="0.0.0.0",
-    port=9389,
-    input_datatype=SDImg2ImgInputs,
+    port=9379,
+    input_datatype=SDInputs,
     output_datatype=SDOutputs,
 )
-@register_statistics(names=["opea_service@image2image_sdwebui"])
-def image2image(input: SDImg2ImgInputs):
+@register_statistics(names=["opea_service@text2image_sdwebui"])
+def text2image(input: SDInputs):
     initialize()
     start = time.time()
-    image = load_image(input.image).convert("RGB")
     prompt = input.prompt
     num_inference_steps = input.num_inference_steps
     guidance_scale = input.guidance_scale
@@ -91,26 +97,12 @@ def image2image(input: SDImg2ImgInputs):
     negative_prompt = input.negative_prompt
     height = input.height
     width = input.width
-    strength = input.strength
 
+    generator = torch.manual_seed(seed)
     images = pipe(
         prompt,
         height=height,
         width=width,
-        num_inference_steps=num_inference_steps,
-        guidance_scale=guidance_scale,
-        generator=generator,
-        negative_prompt=negative_prompt,
-        num_images_per_prompt=num_images_per_prompt,
-    ).images
-
-    generator = torch.manual_seed(seed)
-    images = pipe(
-        image=image,
-        prompt=prompt,
-        height=height,
-        width=width,
-        strength=strength,
         num_inference_steps=num_inference_steps,
         guidance_scale=guidance_scale,
         generator=generator,
@@ -127,20 +119,19 @@ def image2image(input: SDImg2ImgInputs):
             bytes = f.read()
         b64_str = base64.b64encode(bytes).decode()
         results.append(b64_str)
-    statistics_dict["opea_service@image2image_sdwebui"].append_latency(time.time() - start, None)
+    statistics_dict["opea_service@text2image_sdwebui"].append_latency(time.time() - start, None)
     return SDOutputs(images=results)
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("--model_name_or_path", type=str, default="stabilityai/stable-diffusion-xl-refiner-1.0")
+    parser.add_argument("--model_name_or_path", type=str, default="stable-diffusion-v1-5/stable-diffusion-v1-5")
     parser.add_argument("--use_hpu_graphs", default=False, action="store_true")
     parser.add_argument("--device", type=str, default="cpu")
     parser.add_argument("--token", type=str, default=None)
-    parser.add_argument("--seed", type=int, default=42)
     parser.add_argument("--bf16", action="store_true")
 
     args = parser.parse_args()
 
-    logger.info("Image2image server started.")
-    opea_microservices["opea_service@image2image_sdwebui"].start()
+    logger.info("Text2image server started.")
+    opea_microservices["opea_service@text2image_sdwebui"].start()
